feat: Add rebuild + sign workflow
The workflow runs every day at 14:00 UTC, checking if there's new
releases of github.com/rancher/kubectl.

If so, the workflow rebuilds and signs the `rancher/kubectl` image,
pushing it to `kubewarden/kubectl`.

If a `kubewarden/kubectl` image is already present in the repo for the
current tag, the workflow fails early and stops.